+++ health-services/household/src/main/java/org/egov/household/service/HouseholdMemberEnrichmentService.java
@@ -4,9 +4,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.egov.common.models.household.Household;
 import org.egov.common.models.household.HouseholdMember;
 import org.egov.common.models.household.HouseholdMemberBulkRequest;
+import org.egov.common.models.household.Relationship;
 import org.egov.common.service.IdGenService;
 import org.egov.household.repository.HouseholdMemberRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
+import org.springframework.util.ObjectUtils;
 
 import java.lang.reflect.Method;
 import java.util.ArrayList;
@@ -20,6 +23,10 @@ import static org.egov.common.utils.CommonUtils.getIdList;
 import static org.egov.common.utils.CommonUtils.getIdMethod;
 import static org.egov.common.utils.CommonUtils.getIdToObjMap;
 import static org.egov.common.utils.CommonUtils.uuidSupplier;
+import static org.egov.household.Constants.CLIENT_REFERENCE_ID_FIELD;
+import static org.egov.household.Constants.HOUSEHOLD_CLIENT_REFERENCE_ID_FIELD;
+import static org.egov.household.Constants.HOUSEHOLD_ID_FIELD;
+import static org.egov.household.Constants.ID_FIELD;
 import static org.egov.household.utils.CommonUtils.getColumnName;
 
 @Slf4j
@@ -47,6 +54,7 @@ public class HouseholdMemberEnrichmentService {
         enrichHousehold(householdMembers);
         log.info("enriching household members for create request");
         enrichForCreate(householdMembers, uuidList, request.getRequestInfo());
+        enrichRelationshipsForCreate(request, householdMembers);
         log.info("completed enriching household members for create request");
     }
 
@@ -61,13 +69,13 @@ public class HouseholdMemberEnrichmentService {
                 "id", false).getResponse();
         log.info("updating lastModifiedTime and lastModifiedBy");
         enrichForUpdate(hMap, existingHouseholdMembers, beneficiaryRequest);
+        enrichRelationshipsForUpdate(beneficiaryRequest, householdMembers, existingHouseholdMembers);
         log.info("household Members updated successfully.");
     }
 
     public void enrichHousehold(List<HouseholdMember> householdMembers) {
         log.info("getting method for householdId and householdClientReferenceId");
-        Method idMethod = getIdMethod(householdMembers, "householdId",
-                "householdClientReferenceId");
+        Method idMethod = getIdMethod(householdMembers, HOUSEHOLD_ID_FIELD, HOUSEHOLD_CLIENT_REFERENCE_ID_FIELD);
         String columnName = getColumnName(idMethod);
         log.info("getting houseHoldIds for householdMembers");
         List<String> houseHoldIds = getIdList(householdMembers, idMethod);
@@ -82,9 +90,18 @@ public class HouseholdMemberEnrichmentService {
             enrichWithHouseholdId(householdMap, householdMember);
         }
     }
+
     public void delete(List<HouseholdMember> householdMembers,
                        HouseholdMemberBulkRequest beneficiaryRequest) {
         log.info("enriching HouseholdMember with delete information before deletion");
+        if (!CollectionUtils.isEmpty(householdMembers)) {
+            for (HouseholdMember householdMember: householdMembers) {
+                List<Relationship> relationships = householdMember.getMemberRelationships();
+                if (!CollectionUtils.isEmpty(relationships)) {
+                    enrichForDelete(relationships, beneficiaryRequest.getRequestInfo(), true);
+                }
+            }
+        }
         enrichForDelete(householdMembers, beneficiaryRequest.getRequestInfo(), true);
     }
 
@@ -101,5 +118,60 @@ public class HouseholdMemberEnrichmentService {
                 householdMember.getHouseholdClientReferenceId();
         log.info("retrieved householdId");
         return householdId;
+    }
+
+    private static void enrichRelationshipsForCreate(HouseholdMemberBulkRequest request,
+                                                 List<HouseholdMember> validHouseholdMembers) {
+        Method idMethod = getIdMethod(validHouseholdMembers,  CLIENT_REFERENCE_ID_FIELD, ID_FIELD);
+        Map<String, HouseholdMember> householdMemberMap = getIdToObjMap(validHouseholdMembers,idMethod);
+        for (HouseholdMember householdMember : validHouseholdMembers) {
+            List<Relationship> relationships = householdMember.getMemberRelationships();
+            if(CollectionUtils.isEmpty(relationships))
+                continue;
+            List<String> ids = uuidSupplier().apply(relationships.size());
+            enrichForCreate(relationships, ids, request.getRequestInfo(), true);
+            enrichRelationshipsForCreate(request, relationships, householdMember, householdMemberMap);
+        }
+    }
+
+    private static void enrichRelationshipsForCreate(HouseholdMemberBulkRequest request,
+                                                     List<Relationship> relationships, HouseholdMember householdMember, Map<String, HouseholdMember> householdMemberMap) {
+        log.info("enriching relationship for household member: {}", householdMember.getId());
+        List<String> ids = uuidSupplier().apply(relationships.size());
+        enrichForCreate(relationships, ids, request.getRequestInfo(), true);
+        relationships.forEach(relationship -> {
+            relationship.setSelfId(householdMember.getId());
+            relationship.setSelfClientReferenceId(householdMember.getClientReferenceId());
+            relationship.setTenantId(householdMember.getTenantId());
+            if (householdMemberMap.containsKey(relationship.getRelativeClientReferenceId())) {
+                HouseholdMember relative = householdMemberMap.get(relationship.getRelativeClientReferenceId());
+                relationship.setRelativeId(relative.getId());
+            }
+        });
+    }
+
+    private static void enrichRelationshipsForUpdate(HouseholdMemberBulkRequest request, List<HouseholdMember> householdMembers, List<HouseholdMember> existingHouseholdMembers) {
+        log.info("enriching relationships for update");
+        Method idMethod = getIdMethod(householdMembers,  CLIENT_REFERENCE_ID_FIELD, ID_FIELD);
+        Map<String, HouseholdMember> householdMemberMap = getIdToObjMap(householdMembers, idMethod);
+        for (HouseholdMember householdMember : householdMembers) {
+            List<Relationship> resourcesToCreate = new ArrayList<>();
+            List<Relationship> resourcesToUpdate = new ArrayList<>();
+            List<Relationship> updatedResources = householdMember.getMemberRelationships();
+            if(!CollectionUtils.isEmpty(updatedResources)) {
+                resourcesToCreate = householdMember.getMemberRelationships().stream()
+                        .filter(ObjectUtils::isEmpty).toList();
+                resourcesToUpdate = householdMember.getMemberRelationships().stream()
+                        .filter(relationship -> !ObjectUtils.isEmpty(relationship.getId())).toList();
+            }
+
+            if (!CollectionUtils.isEmpty(resourcesToCreate)) {
+                 enrichRelationshipsForCreate(request, resourcesToCreate, householdMember, householdMemberMap);
+            }
+            if (!CollectionUtils.isEmpty(resourcesToUpdate)) {
+                Map<String, Relationship> hmrMap = getIdToObjMap(resourcesToUpdate);
+                enrichForUpdate(hmrMap, resourcesToUpdate, request);
+            }
+        }
     }
 }