+++ health-services/libraries/health-services-models/src/main/java/org/egov/common/models/idgen/IdDispatchRequest.java
@@ -1,0 +1,26 @@
+package org.egov.common.models.idgen;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
+import org.egov.common.contract.request.RequestInfo;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class IdDispatchRequest {
+
+    @JsonProperty("RequestInfo")
+    @NotNull
+    private RequestInfo requestInfo;
+
+    @JsonProperty("IdDispatchRequest")
+    @NotNull
+    @Valid
+    private ClientInfo clientInfo;
+
+
+}